requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
stripe==11.5.0
qrcode[pil]==8.0
python-dotenv==1.1.0
gunicorn==23.0.0
